Separated Execution Engine from Tensor Data Structures
API Changes
===========
In effort to standardize names, I got rid of the more correct method names in favour for standardized names. Also the signatures have changed for those. Here's a table of changes:

+--------------------+------------+
|        OLD         |     NEW    |
+--------------------+------------+
| Trans              |  AddScalar |
| TransInv/TransInvR |  SubScalar |
| Scale              |  MulScalar |
| ScaleInv/ScaleInvR |  DivScalar |
| PowOf/PowOfR       |  PowScalar |
+--------------------+------------+

Where in the past, the `-R` suffix indicate that the scalar value is on the right, the new `-Scalar` methods requires an argument to indicate if the tensor is the left value. This standardization has been fixed in `gorgonia` proper as well. 

New Methods/Exported Functions
==========================
Comparison methods are now exported as methods so you could do `*Dense.ElEq(...)` or `*Dense.EqScalar`

New FMA function which performs fused multiply add.

For more information see https://blog.chewxy.com/2017/09/11/tensor-refactor