API Breaking Changes: 	- Solver now takes []ValueGrad instead of Nodes (which is []*Node)
The reason for this is two fold: The support for BatchNorm operation, which is a verily impure and highly stateful function.
The BatchNorm Op has internal states that need to have their gradients updated as well. But the internal state of BatchNorm isn't really part of
the expression graph, and really it shouldn't be.

The second reason is v0.10.0. In the next version, the solvers will be moved to their own package, as will the other parts.
With this API breaking change, the solver now is less dependent on the other parts of Gorgonia and can be easily separated.