deps: Remove constrained pkg as they were useless
Remove packages which were listed as a "constraint" in Gopkg.toml file
because they are transitive dependencies and if they want to be enforced
they should be listed as "override".

They haven't been swapped to be "override" because they didn't have any
version, revision nor branch, so they were not applying any restriction,
hence didn't make sense to be listed.