Refactor the max/sum ops to share common code. Have the type/inferShape/Do methods behave in a consistent manner:
* Dimensions specified in the "along" parameter are reduced to size 1, but not removed. (Note: this caused TestRepeatOpDoDiff, but this version fixes it.  Perhaps we should make preserving the size-1 dimensions an option of the reduction op?)
* If all dimensions are included, the result will be a scalar.
* If all dimensions but 1 are included, the result is a vector, regardless of which dimension is left intact.

Tests verify that the resulting nodes have the expected shape.

Note: While here, fix a warning on Max's SymDiff where retVal[0] is set when retVal has not been initialized.  I wonder if this is related to gorgonia#323 where SymDiff for StableSoftMax (which uses Max) was failing with a panic (probably not, as the error message there seems unrelated, but probably a good fix anyway).

Closes gorgonia#326