Fixed a bug in the type system wrt transposeOp. When unifying `TensorType{0, a}` with a type constant (e.g. `Float64`), the `TensorType` will return `[a]` in `.Types()`.
Since we define `TensorType{0, a}` as equivalent to `a`, something
must be done.

There are a few ways we can solve this problem:

1. We can make the type constants (e.g. `Float64`) return itself when
asked for the smaller types
2. We can canonicalize `TensorType` such that when the `Dims` == 0, it
returns the `.Of` instead.

1 is problematic - type constants are atomic! They are not made of
anything smaller (literally the meaning of the word "atomic"). They're
not recursive types either. Changing that would have a huge amount of
implications.

So a new method, `.Canonical()` is added to the
`TensorType`.

As well as that, `NewFunc` was added to the type system such that
canonicalization is handled before a new `*hm.FunctionType` is
created.