This seems to fix the no compiling errors
So when I try to run the code, it wont compile. It looks like the interface `Memory` in package `gorgonia.tensor` is supposed to declare a function `Pointer` (or at least this package expects it to). When I added that function to the interface it broke many things so i have opted to do this instead. I'm pretty sure it wont break anything