Fix BatchedMatMul bugs (#555)
* Fix batchedMatMul bug when applying to 1x1 matrices

The operator batchedMatMul slices the batch of matrices batch by batch. But the slice of tensor will reshaped by tensor.Slice operation to remove those dimension with 1-shape. For example, for 10x1x1 tensor A to multiply 10x1x10 matrix B. The slice of A will be reshaped to scalar, which results in error to multiplying matrix B's slice which has 1x10 shape.

* Fix batchedMatMul bug when applying to 1x1 matrices

The operator batchedMatMul slices the batch of matrices batch by batch. But the slice of tensor will reshaped by tensor.Slice operation to remove those dimension with 1-shape. For example, for 10x1x1 tensor A to multiply 10x1x10 matrix B. The slice of A will be reshaped to scalar, which results in error to multiplying matrix B's slice which has 1x10 shape.

* Remove the duplicate code caused by incorrect merge for BatchedMatMul

* Remove reshape for result matrix in BatchedMatMul if preallocated

* Fix bug in re-using slice's underlying data for newShape in batchedMatMul function

* Add more unit tests for BatchedMatMul to increase coverage